Initial commit
At this point, tvw-sdr performs the same init that the closed driver
does, and in less than a second (compared to over 45 seconds).  The main
differences in register read-outs are from register 0x4400.  There is
basic I2C read support in place, which is enough to read the 16 bytes of
non-extended register space in the TDA18271 tuner behind the Theater IC.

License under GPLv2+.